api/requirements.txt
# Hyperion Fleet Manager API -- production dependencies
# Pin major.minor to avoid breaking changes while allowing patches.

# --- Web framework ---
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0

# --- Data validation & settings ---
pydantic[email]>=2.10.0,<3.0.0
pydantic-settings>=2.7.0,<3.0.0

# --- AWS ---
boto3>=1.35.0,<2.0.0
aioboto3>=13.0.0,<14.0.0

# --- Data stores ---
redis[hiredis]>=5.2.0,<6.0.0
celery[redis]>=5.4.0,<6.0.0

# --- Authentication ---
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0

# --- Observability ---
structlog>=24.4.0,<25.0.0
python-json-logger>=2.0.0,<3.0.0

# --- HTTP client (used in tests and service-to-service calls) ---
httpx>=0.27.0,<1.0.0
